+++ packages/nocodb/src/models/index.ts
@@ -40,3 +40,4 @@ export { default as View } from './View';
 export { default as LinksColumn } from './LinksColumn';
 export { default as Notification } from './Notification';
 export { default as PresignedUrl } from './PresignedUrl';
+export { default as UserRefreshToken } from './UserRefreshToken';
